qt: support 2d surfaces for horizontal scroll
Fixes #20498

To avoid detection of erratic movements along the lines
introduce allowance zone for each axis.

Angle determines how close to axis angle delta is.
For horizontal events it is set to 30 degrees (15 for each side).
For vertical events angle is 53 degrees (30 * 16/9).
It is a little wider since touchpads are asymmetrical.

Signed-off-by: Pierre Lamot <pierre@videolabs.io>