opengl: fix missing flag initialization
The flag filter_planes was not explicitly initialized, and non-plane
filters assume that it is set to false by default.

It still worked because the filter happens to be initialized by calloc()
in the end, but we shouldn't rely on this detail (and the other fields
are also explicitly initialized).

Signed-off-by: Alexandre Janniaux <ajanni@videolabs.io>