libplacebo: move vout to API-agnostic module
Rebases vulkan/display.c ontop of the new API-agnostic libplacebo GPU
abstraction module. This opens up libplacebo/display.c to work with API
backends other than Vulkan.

Note: this has the side effect of bumping up the minimum libplacebo
dependency, because now the libplacebo vout is always compiled when
libplacebo is available. On the plus side, however, it simplifies the
configure check. That said, 1.7.0 is from 2018, and packaged everywhere
(including debian stable). so this should be fine.