libplacebo: add MakeCurrent/ReleaseCurrent wrappers
Needed for OpenGL API support, at least on versions of libplacebo prior
to v3.128. It's worth pointing out that this logic is therefore only
temporary - once the libplacebo version is bumped past v3.128, this can
be removed again.

This commit only adds the API, with the OpenGL implementation coming in
a following commit.