spu(audiobargraph,logo): fix redundant align attribute
i believe that this is an entirely non-functional change upon the
belief that this property is not used in absolute mode (and thus is
actually rather pointlessly being set). i'm just changing it out of
pedantic correctness.

the top-right setting did not match the positioning done by absolute
mode (see `SpuRegionPlace()` in vout_subpictures.c), and was
inconsistent with what was set in marq and rss plugins.

this originates for logo with 5a33cfd28d051ecd885bbcbee0e3465fba8ae1a7
whereby the author decided for no apparent reason to change the
property to reflect the previous position default of 6 (top-right).

this seems to have then been inherited through copy & paste i guess
into the creation of audiobargraph_v in 0e3ec2274b60aca41c60c2dc37d61e408be70236.