cmdline: reorganise error handling
...to prepare for text translation and colour labels.

(i am not certain that tacking the option onto the end, as we were,
will be suitable for all translations).

i've avoided using allocation to construct the option name portion of the
output, since i expect that would not be liked, and i'm not convinced about
it myself. it can always be reconsidered later however.