cmdline: use cleaner 'see --help' text
old: "Try `vlc --help' for more information."
new: "For more information try `--help`"

in experiments, the new text proved much cleaner. it becomes yet more
clean, having dropped "vlc" from `vlc --help`, when we then drop the
quotes around the remaining `--help` in the commit after next.

note, we now no longer have "vlc" in any of the error output. this
is not a problem for the vlc app itself obviously; i hope it is not
a problem for any other libvlc use cases.