vulkan: rename instance.h to platform.h
With the removal of the actual vulkan instance from this abstraction in
883a865a, calling this "instance" is confusing and misleading. It should
be called "platform", because the only thing the abstraction is designed
to contain is platform-specific state and operations.

Rename it, and the associated type, for clarity. This frees up the name
`vlc_vk_t` to be used for an actual Vulkan instance, if we ever decide
we need one in the future. (e.g. for GPU filters)