libplacebo: add support for BT.2390 EETF
This has already been the default in most libplacebo versions, in that
`pl_color_map_default_params.tone_mapping_algo` was equal to this value.
Despite this, it was never exposed in the UI, meaning users could not
actually manually select it.

Update the string to make it clear that this is the intended/recommended
default.