libplacebo: fix use-after-free on uninit
`pl->ops->close` destroys the GPU instance, after which the swapchain is
undefined. Fix it by moving the swapchain destruction to `ops->close`
itself, where it can be done in the correct order.

(This also fixes a bug where, technically, the opengl instance attempted
destroying the swapchain while the context was not current on the
thread)