libplacebo: add support for custom LUTs
Unfortunately the UI for loading these is a bit clunky as they're hidden
behind so many layers of menu and advanced options, plus the fact that
the new LUTs still only get applied on vout reinit.

But it's still a feature that a lot of users have requested heavily, so
supporting them at all is a good thing. The implementation is rather
straightforward even though the amount of #ifdef's are really getting on
my nerves.

Can somebody just time travel me into a future universe in which I can
make libplacebo v4 the minimum dependency?