http: chunked: fix chunked transfer encoding output
Chunks in chunked transfer-encoding are supposed to be ended with \r\n.
Without that, servers can be confused on how to parse the next chunk
size, and are typically returning a 400 Bad Request. This can be seen in
the Mozilla developer documentation[1], or more exhaustively in the
document RFC[2] section 4.1.

This is in particular needed to signal that no other headers are exposed
in the chunk.

[1]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Transfer-Encoding#chunked_encoding
[2]: https://datatracker.ietf.org/doc/html/rfc7230#section-4.1