qt: (hotkeys) fix "Unset" matching issue
the `locale` boolean param determines whether or not we are asking for a
translation to be performed. in the "unset" (error?) case, translation was
incorrectly always being performed upon the string "Unset", which meant
that any caller not working with translations and wanting to check for the
"Unset" condition would fail to work correctly in situations where the
translation would result in something other than exactly "Unset".

if translation has not been requested, return the untranslated form.

this means that now `KeySelectorControl::selectKey()` when reacting to
`QDialog::Accepted` will correctly store the untranslated form of "Unset"
where applicable in the data attribute. (the data attribute is supposed to
always hold an untranslated copy).

this also means that the data attribute comparison within the
`KeyInputDialog::checkForConflicts()` function will now correctly work in
translation situations.

note: it is unclear why we return "Unset" here rather than an empty string,
creating a distinction from all other unset actions. git history shows that
there was a shift from using the text "Unset" to using an empty string, yet
this was left, and without any clear explanation as to why, with a very
real possibility that it was simply overlooked. here we just leave that
question alone, to return to later.