qt: (hotkeys) avoid re-doing key conversions
we had:
 1. event making translated copy for 'selected' text.
 2. conflict check making an untranslated copy (one commit back in this
    set it was another translated copy).
 3. the 'accepted' code, if triggered, making a further one of each.

a single copy of each can be created and stored in public attributes by
the event handlers, and used in all of these places, without any negative
impact. this replaces the need for the public `keyValue` attribute, and
replaces the need to give the conflict checker a copy of the integer.