qt: (hotkeys) fix missing translation of initial displayed global assignment
the displayed text is always displayed translated for non-global, and this
is the case for edits for global, but the initial unedited displayed text
was not translated.

due to the way the global data is temporarily stored in a `QMap`, it seems
that the easiest way to fix this is to simply convert from `QString` to
raw string pointer with `qtu()` and then back to `QString` with translation
applied with `qfut()`.