qt: (hotkeys) fix overlooked old use of "Unset" for an unset item situation
the "Unset" text used to be used for all unset cases, but a long time ago
all except this one got changed to use an empty string instead. having
spent a bit of time trying to track down an answer in the git history as to
the question of why, i gave up. i found no clear answer in the time spent
looking. it very well may have been simply overlooked rather than
deliberately left to create a clear distinction that a failure occurred.

i'm going to assume that it was overlooked. i've thus changed it to return
an empty string. i don't expect this to cause any problems for the two
different uses in the Qt code.