qt: (hotkeys) remove redundant problem check
we now rely upon checking option name rather than shorttext. it is somewhat
less likely that a mistake with duplicate option names will have been made
compared to duplicate shorttext, and duplicate option names presents an
even bigger and broader problem that we currently do nothing about. this is
not very useful towards helping with that, we'd want a full check of the
entire table's option names if we cared to properly check things here.