qt: (hotkeys) do not skip item on null shorttext (action label)
unlikely and better to add it rather than skip. being there with a blank
label would better help diagnose what went wrong in testing, or if it
slipped past testing, someone is more likely to notice and report the
problem leading to a fix.

note that we must conditionally use `qfut()` because the underlying
translation function does not like nulls, and my fix for that was rejected.