config: use common function for identifying 'general' subcats
...removing the code duplication and helping avoid mistakes.

this reverts the fix in 08134b3f77fad1a0da3a20ddc869aadb8fe997cc
that stopped treating the 'advanced-misc' subcat as a general one
for the macos interface module, as a means of solving some issues
with displaying nodes for plugins associated with the 'advanced'
category. later commits properly fix the issues with the macos
tree construction code. in the meantime i want the predicate that
every cat has a general subcat restored such that it can be
properly relied upon once again by further preference tree code
improvements.