qt: improve a subcat ID init
(non-functional).

a value of `0` in some parts of the tree construction code corresponds
to unknown/unset/invalid, for which the new 'unknown' defines are used.
a value of `-1` corresponds to 'hidden'.

this instance of `-1` was not changed to use the `SUBCAT_HIDDEN` define
when that was introduced because it was not clear that 'hidden' truly
represented what was really intended by this init value and in fact
that the wrong init value was likely being used here.

since this attribute is only ever used for `TYPE_CATSUBCAT` instances,
and is always overwritten by the actual subcat ID upon creation, it
really doesn't matter what it is initialised with. to be more correct,
clear and consistent though, let's init as 'unknown' - the correct
define for initialisation purposes.