iso639: disable some unnecessary defs and simplify a few others
-- deletions: --

except for the "Artificial" entry, all of those disabled are for ancient
languages. i expect that users are highly unlikely to encounter subtitles
with any of these languages, and thus it is best to disable them.

they have been disabled rather than deleted to try to minimise difficulties
for future bulk updates from the glibc set.

-- simplifications: --

 - the "Interlingua" entry was considered excessively large in review.

 - the Greek and Occitan/Provençal entries do not really need to mention
   the dates, since we've disabled the entries for ancient forms of those
   languages and thus the clear distinction is unnecessary.

 - the Low German/Saxon entry did not really need the "German/Saxon, Low"
   form included.

 - the Church Slavic entry, according to wikipedia ([1], [2]), seems to
   cover both the ancient ("old") form (9th-11th century) and newer
   (non-"old") forms. we have no interest in the ancient form, and we
   would disable that if it had its own entry. my understanding (please
   forgive my ignorance) from [1] is that the non-"old" form is still in
   use and so still relevant. certainly the entry already existed in vlc.
   it's excessively long and a request to shorten it was made in review,
   so here i've just simplified the label by removing the unnecessary
   references to the "old" forms.

fwiw, in the Greek, Occitan/Provençal and Interlingua cases we happen to
restore the labels back to what they were before we did the update. the
Church Slavic entry is slightly different.

the "simplified label" comments have been left for the benefit of anyone
performing a further bulk update in the future, to note the deliberate
deviations from the glibc entry text.

[1]: https://en.wikipedia.org/wiki/Church_Slavonic
[2]: https://en.wikipedia.org/wiki/Old_Church_Slavonic