d3d11: remove Query based rendering wait
It is imprecise and on seek it may wait a lot of time as, while it's waiting,
the decoder threads will flood new frames, delaying that much longer until the
ID3D11DeviceContext is finished doing things. This workaround did more harm
than good.

Now that we have ID3D11Fence support (any recent Win10) the query based system
should not be used. For older Windows using D3D11 we'll have less precise
estimation of the time it took to actually render, meaning the frame drop will
be less effective.