iso639: de-duplicate code lookup handling
...with a new function, which also replaces direct use of the three
code-specific functions.

all users of these three functions had a common pattern of using the
iso-639-1 lookup for len==2 codes, and for len==3 doing a 2B lookup,
falling back to a 2T lookup if that failed.

this pattern is now implemented in the new `vlc_find_iso639()` function.

the three individual functions are now made private, not being needed
publicly anymore. they also now just do a straight `strcmp()` rather than
`strncmp()` because they are only called for the right length of strings.

the implementation will be tweaked slightly in the next commit.