iso639: add capability for 'any-field' alternate searches
esout, dvdnav and bluray are currently using a very old hack of searching
all fields of the raw language table for searches where they want to allow
matches against the names, not just codes. this is a step towards getting
rid of that ugly old hack by offering proper functionality for doing such
a search.

the implementation currently duplicates that used by the three previously
mentioned use cases. i intend to rework the implementation in a subsequent
commit; first though we need to just set things up for simply removing the
hack. that new implementation will better justify this being done with a
flag rather than a separate function.