mux/ts: ditch unnecessary helper function
this was introduced in b8774b3c1a183e41dc8b9e7be55e48cc10ddac91 as part of
de-duplicating code, however the two blocks that used the relevant code
were also merged in the same commit, making the function unnecessary.

this is being done partly as cleanup before the next commit, which would
remove it anyway, but doing it separately first makes the diffs more
clear.

this brings the use of the lookup functions back into line with use
everywhere else in the codebase.