esout: switch from using iso-639-1 to iso-639-2B codes
...for es language attributes.

while all iso-639 records in the table (iso-639_def.h) currently have a
two char iso-639-1 code, with the exception of "Original audio", if/when
MR 146 ([1]) is merged, there will be many records with no iso-639-1 code,
which will cause empty strings to be captured in the language properties
for such languages.

all records, now and after MR 146, have an iso-639-2B code so switching to
that will solve the problem. (all but one have iso-639-2T fwiw).

[1]: https://code.videolan.org/videolan/vlc/-/merge_requests/146