configure.ac: remove -framework,AppKit
Those frameworks only exist for MacOSX, so it was invalid to set them up
in the general case. Fortunately, it wasn't really used on iOS and tvOS
for the executables so it didn't have consequences.

On MacOSX, the targets needing this framework already link it so it is
not needed.