extras/tools: libtool: fix support for bitcode
When building with bitcode enabled (*FLAGS+=-fembed-bitcode), the
following happens:

 - with LDFLAGS += -fembed-bitcode, the vanilla libtool version will
   remove the -fembed-bitcode flag, being an unknown flag, so bitcode
   won't actually be enabled, and final link steps might complain that
   the dylibs (libvlccore.dylib for instance) doesn't actually have
   bitcode embedded.

 - once fixed (either by this patch or by using -Wl,-fembed-bitcode),
   the plugins will be compiled with -module, which default to using
   MH_BUNDLE on iOS/tvOS/MacOSX instead of MH_DYLIB. Indeed, dylibs can
   be dlopened only since MacOSX 10.3/10.4, and can be dlclosed only
   since MacOSX 10.5, so MH_BUNDLE was the primary target for libtool
   module support. However, MH_BUNDLE is not compatible with the bitcode
   support from Clang.

Since we support MacOSX 10.11 as minimum, we can default to using
MH_DYLIB for plugins to have the bitcode support.

This new libtool patch add the explicit support for -fembed-bitcode in
libtool --mode=link flags, and change the flags used when using -module
in libtool to switch to dylibs.