Makefile.am: use a macro to set rpath
Move the use of -rpath '$(foodir)' into a macro that can be
conditionnally defined. It allows the variable to be conditionally
removed as it will be the case for static builds where this does not
make sense, just like the automake integration is doing.

The addition of -rpath '$(foodir)' is only needed when automake cannot
infer it, ie. when the EXTRA_LTLIBRARIES and $(LTfoorbar_plugin)
substituation mechanisms are used instead of automake conditionals.

Check with the following command for completness status:

        $ git grep '\-rpath' | grep -v _RPATH

No functional changes.

Refs #25156