Makefile.am: add redirection level for plugins
The plugins targets are directly pushed to category_LTLIBRARIES which
would immediately set up the -rpath argument indicating where the
library should be installed.

When doing partial linking, the libraries we'd like to install are not
those targets but the partially linked ones, so add an indirection level
to redirect them to noinst_LTLIBRARIES.

This indirection also allows us to proceed with specific post-process
for each object and generating the targets for the real final artifacts.

The completeness of the work can be checked against the following
command:

        $ git grep _LTLIBRARIES \
        | grep -v EXTRA \
        | grep -v PLUGINS \
        | grep -v noinst_ \
        | grep -v pkglib_ \
        | grep -v check_ \
        | grep -v -E '=$' \
        | grep -v contrib/

Refs #25156