uikit-inhibit: fix threading issue
The UIKit functions must be called from the main thread, whereas the
functions from the inhibiter can be called from the vout thread. It
didn't seem to have any consequence on iPad, or the testing I did was
not triggering the crash, but it's leading to a crash on tvOS when
running with VLCKit.