deinterlace: add --deinterlace-filter option
The option --deinterlace-mode selects the deinterlace algorithm to use,
but the same algorithm might be implemented by several modules, for
example a CPU version and an OpenGL version.

For that purpose, add a separate option to select the module(s) to use.
The default value is "deinterlace".

To select the blend deinterlace algorithm:

    --deinterlace-mode=blend

or equivalently ("deinterlace" is the default filter):

    --deinterlace-filter=deinterlace --deinterlace-mode=blend

To run the OpenGL version (the mode is "auto"):

    --deinterlace-filter=glblend

The module glblend only implements the "blend" algorithm, it can be
passed explicitly:

    --deinterlace-filter=glblend --deinterlace-mode=blend

To pass a list a modules, executed in order until one works:

    --deinterlace-filter=glblend,deinterlace --deinterlace-mode=blend