medialib: acquire priority access
During indexation, the medialibrary executes a lot of write SQL queries
from a background thread. As a consequence, requests initiated by VLC to
retrieve data to show in the UI are excessively delayed (see [1]).

These requests are not executed on the UI thread anymore (since [2]),
but it is still not acceptable to wait tens of seconds just to display
local data.

Therefore, a new priority access mechanism has been implemented in
medialibrary ([3]). Acquire this priority access for all Get and List
requests initiated from VLC.

[1]: https://code.videolan.org/videolan/vlc/-/issues/22687#note_259820
[2]: https://mailman.videolan.org/pipermail/vlc-devel/2020-November/140488.html
[3]: https://code.videolan.org/videolan/medialibrary/-/merge_requests/375