qt: use strongly-typed connects
This replaces almost all use of `SIGNAL()` and `SLOT()` with Qt5-style
strongly-typed connections. The `CONNECT()`, and `DCONNECT()` helpers
built around them have been removed.

Cases involving the `BUTTON_*` macros will be tackled in a separate
commit to improve clarity.

A few cases remain using `SIGNAL()` and `SLOT()` for now due to
complications that should be dealt with separately.