demux/subtitle: use only filename for filename language substr extraction
...and thus prevent some ugly failures.

(the string given to the function is the full filepath not just the filename).

the attempt to determine the language from subtitle filenames is based upon
a single common pattern - PATH/filename.LANG.ext. whilst it works just fine
for this pattern, it is not the only pattern commonly used, for instance
PATH/Subs/x_LANG.ext (where 'x' is an integer).

in such cases where the period for the extension is the only one in the
filename, the function could produce an ugly result should any directory in
the path happen to contain a period (if not, NULL would be returned). it
would incorrectly capture a chunk of the path as part of the substring
extraction, producing results like "FOOBAR/Subs/1_English" (or worse) which
then end up as the language name displayed under the subtitle menu and
elsewhere.

this commit strips the string processed down to filename only and thus
prevents such ugliness. the next commit will introduce proper handling for
the just mentioned alternate common pattern.

(cherry picked from commit 0984f11c58865af87c2b73676f12120702af25b3)