demux/subtitle: remove pointless statement
we find the period for the extension and set it to null such that we can
then find the second to last period that should preceed the substring we
want to extract. there is absolutely no point in restoring the period
afterwards in the working copy which we are then just about to destroy.

the comment did not even make any sense.