demux/subtitle: handle PATH/Subs/1_English.srt type filename lang extraction
the only pattern handled was PATH/filename.LANG.ext. another common one is
PATH/Subs/x_LANG.ext which this adds handling for.

this simply replies upon falling back to trying to get the substring after
the last underscore if trying to get the substring after a period fails.

we do not explicitly require the second pattern to only occur in files
found under a 'Subs' subdir, since it is not certain that there is value in
implementing such a restriction.