demux/subtitle: prepare for lang detection via codec properties
at least one subtitle format handled by this demuxer may hold the language
as a property specified within the file. we should allow the parser to
extract and use that as an alternative to the filename based substring
extraction. this sets things up to allow the parser functions to provide
that extracted property string.