demux/subtitles: clarify debug message
the substring obtained from filename extraction in some cases is perfect
but in other cases may not be a language at all, just some portion of the
filename. stating 'detected language FOO' is a bit odd if it turns out to
not actually be a language name that we've extracted. let's fix that by
clarifying what we've actually retrieved, and thus distinguish the less
reliable filename extraction result from the likely more reliable property
available in some subtitle files.

also, enclose in quotes in both cases. for the filename based case since
this simply makes sense. in the property case, since this may be a language
code, as it is for ASS/SSA.