macosx: Fix tiny bugs in volume control
Fix a problem in the parsing string used for the volume control tooltip.
The old string formatted a float as an int, leading to printing wrong
values.

Additionally, set the default value of the volume on startup to 1 (the
correct value), instead of 0.

Clean the LibraryWindow xib from 3 old variables, that were removed from
the class, but not the xib.